pyproject.toml
[tool.poetry]
name = "steganoproject"
version = "0.0.1"
description = "Steganography Project"
authors = ["Ahmet Kucukdag <akucukdag@etu.edu.tr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
cryptography = "^41.0.7"
setuptools = "^69.0.2"
wheel = "^0.42.0"
Pillow = "^10.1.0"
pyinstaller = "^6.2.0"
numpy = "^1.26.2"
matplotlib = "^3.8.2"
crypto = "^1.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
